Add result set mutation to SQLLab.
